Cyber-Security-Swiss-Army-Knife/network_probe/requirements.txt
scapy
python-nmap
pyfiglet